Moved download_report test file into analysis folder Added test that if HTML/CSV report download is not allowed - it is not available Added download validation for HTML/CSV reports for existing analysis tests Cleaned up unused imports in some tests
